Move custom exception para um arquivo de exceções da app
