+++ covid19/tests/test_spreadsheet_validator.py
@@ -11,10 +11,10 @@ from django.test import TestCase
 
 from brazil_data.cities import get_city_info
 from core.models import Table
+from covid19.exceptions import SpreadsheetValidationErrors
 from covid19.models import StateSpreadsheet
 from covid19.spreadsheet_validator import (
     format_spreadsheet_rows_as_dict,
-    SpreadsheetValidationErrors,
     validate_historical_data
 )
 from covid19.tests.utils import Covid19DatasetTestCase
